Made another component. Some errors to fix
