+++ src/components/ProductsSearch.tsx
@@ -6,6 +6,9 @@ import { FaMagnifyingGlass } from "react-icons/fa6";
 import { Listbox, Transition } from '@headlessui/react'
 import clsx from "clsx";
 import { SortOrders } from "@/types";
+import { getCategories } from "@/lib/data";
+import CategoryDropdown from "./CategoryDropdown";
+
 
 const sortableKeys = [
   { label: "Date Added", value: "created_at" },
@@ -19,13 +22,14 @@ const orderKeys = [
 
 const categoryKeys = [
   { label: "All Categories", value: "" },
-  { label: "Candles", value: "d0e73609-8c23-491f-8044-361e3bad95ba02ef648f-f4a5-4852-b3fd-0cc17e8cc2999c152eac-515f-4429-b610-65ff1f4624a004e5ea35-2663-4cb4-bff2-cbd8b0b387eb" },
-  { label: "Jewelry", value: "empty" },
-  { label: "Pottery", value: "empty" },
-  { label: "Textiles", value: "empty" },
-  { label: "Paintings", value: "empty" },
-  { label: "Music", value: "0c44b8a7-184e-4644-bfed-98b0cf30f674" },
+  { label: "Candles", value: "Candles" },
+  { label: "Jewelry", value: "Jewelry" },
+  { label: "Pottery", value: "Pottery" },
+  { label: "Textiles", value: "Textiles" },
+  { label: "Paintings", value: "Paintings" },
+  { label: "Music", value: "Music" },
 ]
+
 
 export default function ProductsSearch() {
   const searchParams = useSearchParams();
@@ -44,10 +48,6 @@ export default function ProductsSearch() {
     const params = new URLSearchParams(searchParams);
     return params.get('order') || SortOrders.Descending;
   });
-  const [category, setCategory] = useState(() => {
-    const params = new URLSearchParams(searchParams);
-    return params.get('category') || "";
-  });
 
   const updateSort = useCallback((newSort: string) => {
     setSort(newSort);
@@ -65,13 +65,6 @@ export default function ProductsSearch() {
     replace(`${pathname}?${params.toString()}`);
   }, [setOrder, searchParams, pathname, replace]);
 
-  const updateCategory = useCallback((newCategory: string) => {
-    setCategory(newCategory);
-    const params = new URLSearchParams(searchParams);
-    params.set('category', newCategory);
-
-    replace(`${pathname}?${params.toString()}`);
-  }, [setSort, searchParams, pathname, replace]);
 
   const updateSearchTerm = useCallback((event: React.ChangeEvent<HTMLInputElement>) => {
     setSearch(event.target.value);
@@ -96,9 +89,6 @@ export default function ProductsSearch() {
     return Object.entries(orderKeys).find(([_, { value }]) => value === order)?.[1]?.label || "";
   }, [order]);
 
-  const categoryLabel = useMemo(() => {
-    return Object.entries(categoryKeys).find(([_, { value }]) => value === category)?.[1]?.label || "";
-  }, [category]);
 
   return (
     <div>
@@ -212,50 +202,7 @@ export default function ProductsSearch() {
           </Listbox>
         </div>
         <div className="flex-1">
-          <Listbox value={category} onChange={updateCategory}>
-            {({ open }) => (
-              <>
-                <div className="relative mt-2">
-                  <Listbox.Button className="relative w-full cursor-default rounded-md bg-white py-1.5 pl-3 pr-10 text-left text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 focus:outline-none focus:ring-2 focus:ring-indigo-500 sm:text-sm sm:leading-6">
-                    <span>{categoryLabel}</span>
-                  </Listbox.Button>
-
-                  <Transition
-                    show={open}
-                    as={Fragment}
-                    leave="transition ease-in duration-100"
-                    leaveFrom="opacity-100"
-                    leaveTo="opacity-0"
-                  >
-                    <Listbox.Options className="absolute z-10 mt-1 max-h-56 w-full overflow-auto rounded-md bg-white py-1 text-base shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none sm:text-sm">
-                      {categoryKeys.map((key) => (
-                        <Listbox.Option
-                          key={key.value}
-                          className={({ active }) =>
-                            clsx(
-                              active ? 'bg-indigo-600 text-white' : 'text-gray-900',
-                              'relative cursor-default select-none py-2 pl-3 pr-9'
-                            )
-                          }
-                          value={key.value}
-                        >
-                          {({ selected, active }) => (
-                            <div className="flex items-center">
-                              <span
-                                className={clsx(selected ? 'font-semibold' : 'font-normal', 'ml-3 block truncate')}
-                              >
-                                {key.label}
-                              </span>
-                            </div>
-                          )}
-                        </Listbox.Option>
-                      ))}
-                    </Listbox.Options>
-                  </Transition>
-                </div>
-              </>
-            )}
-          </Listbox>
+          <CategoryDropdown />
         </div>
       </div>
     </div>
